Remove option to edit the default theme
If the system facilitates in updating the default theme, it is to be expected that the changes can be kept. But since the default theme is under version control, it can be overwritten with an update. Therefore it is better for remove the option to edit the default theme.

This fixes a part of #175